+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 - Support for new all-Contentful content into the locale links.
 - Support for new all-Contentful content into the localized anchor links.
 - Support for new all-Contentful content into the alternative links in the head for each page.
+- Support for search when using Contentful.
 
 ### Changed